feat: move log type into log
