+++ src/entity/index.ts
@@ -3,5 +3,4 @@ export * from "./log";
 export * from "./category";
 export * from "./post";
 export * from "./tag";
-export * from "./logType";
 export * from "./logFactory";
