--- src/entity/logType.ts
@@ -1,10 +1,0 @@
-import { registerEnumType } from "type-graphql";
-
-export enum LogType {
-  LOGIN = "LOGIN",
-  POSTING = "POSTING",
-}
-
-registerEnumType(LogType, {
-  name: "LogType",
-});
